refactor: split start test support modules
